Restructure types placement
