--- src/services/board/board.model.d.ts
@@ -1,11 +1,0 @@
-export interface ColumnInterface {
-
-}
-
-export interface BoardInterface {
-  id: string;
-
-  title: string;
-
-  columns: ColumnType[];
-}
