--- src/services/task/task.model.d.ts
@@ -1,15 +1,0 @@
-export interface TaskInterface {
-  id: string;
-
-  title: string;
-
-  description: string;
-
-  order: number;
-
-  userId?: string | null;
-
-  boardId?: string | null;
-
-  columnId?: string | null;
-}
