--- src/services/users/user.model.d.ts
@@ -1,9 +1,0 @@
-export interface UserInterface {
-  name: string;
-
-  id: string;
-
-  login: string;
-
-  password: string;
-}
